+++ server/web/controller/model/config.go
@@ -1,6 +1,6 @@
 /**
  *
- * (c) Copyright Ascensio System SIA 2025
+ * (c) Copyright Ascensio System SIA 2026
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -54,7 +54,10 @@ type User struct {
 type Customization struct {
 	Goback  Goback `json:"goback"`
 	UiTheme string `json:"uiTheme,omitempty"`
+	Chat    bool   `json:"chat"`
 	Close   Close  `json:"close,omitempty"`
+	Plugins bool   `json:"plugins"`
+	Macros  bool   `json:"macros"`
 }
 
 type Close struct {