+++ webapp/src/redux/actions/index.ts
@@ -3,7 +3,7 @@
 
 /**
  *
- * (c) Copyright Ascensio System SIA 2025
+ * (c) Copyright Ascensio System SIA 2026
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -23,3 +23,4 @@ export {closePermissions, openPermissions} from './permissions';
 export {closeEditor, openEditor} from './editor';
 export {closeManager, openManager} from './manager';
 export {closeConverter, openConverter} from './converter';
+export {updateHealthStatus} from './health';