+++ webapp/src/redux/reducers/index.ts
@@ -3,7 +3,7 @@
 
 /**
  *
- * (c) Copyright Ascensio System SIA 2025
+ * (c) Copyright Ascensio System SIA 2026
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -23,6 +23,7 @@ import {combineReducers} from 'redux';
 
 import {converterModal} from './converter';
 import {editorModal} from './editor';
+import healthReducer from './health';
 import {managerModal} from './manager';
 import {permissionsModal} from './permissions';
 
@@ -31,4 +32,5 @@ export default combineReducers({
     editorModal,
     managerModal,
     converterModal,
+    health: healthReducer,
 });